+++ src/main/java/ru/kata/spring/boot_security/demo/dao/UserRepository.java
@@ -11,9 +11,9 @@ import java.util.Optional;
 public interface UserRepository extends JpaRepository<User, Integer> {
     Optional<User> findByName(String Name);
     User findByEmail(String email);
-    Optional<User> findById(Integer id);
+    Optional<User> findById(Long id);
     <S extends User> S save(S entity);
     List<User> findAll();
     void delete(User user);
-    void deleteById(Integer integer);
+    void deleteById(Long integer);
 }
